==> IND/Code/2_predict.ipynb <==
# LB:(20240608_miyaki_5.json) 

import os 
import json
import warnings
import numpy as np
import pandas as pd
import multiprocessing
import lightgbm as lgb
import xgboost as xgb
from sklearn.ensemble import RandomForestClassifier
from catboost import CatBoost,Pool

import torch
from pytorch_tabnet.pretraining import TabNetPretrainer
from pytorch_tabnet.tab_model import TabNetClassifier

import tensorflow as tf
from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import Dense, Dropout, BatchNormalization
from tensorflow.keras.optimizers import Adam
from tensorflow.keras.losses import BinaryCrossentropy
from tensorflow.keras.metrics import AUC
from tensorflow.keras.callbacks import EarlyStopping

from sklearn.decomposition import PCA
from sklearn.preprocessing import StandardScaler
from sklearn.model_selection import train_test_split, StratifiedKFold, GroupShuffleSplit
from sklearn.metrics import recall_score, precision_score, roc_auc_score,confusion_matrix

CORE_NUM = multiprocessing.cpu_count()
device = "cuda" if torch.cuda.is_available() else "cpu"
warnings.simplefilter('ignore')
os.makedirs("../submit/", exist_ok=True)

print(f'multi:{CORE_NUM} | device:{device}')

df_master = pd.read_parquet('../test_data/cleaned_pid_to_info_all_v6.parquet')
df_master = df_master.rename(columns={'id':'paper_id'})
df_train_master = pd.read_parquet('../test_data/train_author.parquet')
df_test_master = pd.read_parquet('../test_data/ind_test_author_filter_public.parquet')
print(f'train:{df_train_master.shape} | test:{df_test_master.shape}')

def get_ids(df, proc):
    merge_keys = ['author_id','paper_id','label']
    if proc == 'train':
        df_normal = df.explode('normal_data')[['id','normal_data']]
        df_normal.columns = ['id','paper_id']
        df_normal['label'] = 1
        df_outliers = df.explode('outliers')[['id','outliers']]
        df_outliers.columns = ['id','paper_id']
        df_outliers['label'] = 0
        df = pd.concat([df_normal,df_outliers])
        df.columns = ['author_id','paper_id','label']
    elif proc == 'test':
        df = df.explode('papers')[['id','papers']]
        df['label'] = 0
        df.columns = ['author_id','paper_id','label']
    return df[merge_keys]
    
def get_feature(df, proc):
    merge_keys = ['author_id','paper_id','label']
    
    df_glm = pd.read_parquet(f'../test_feature/{proc}_glm.parquet')
    df = df.merge(df_glm, on=merge_keys, how='left')
    
    df_sci = pd.read_parquet(f'../test_feature/{proc}_scibert_nli.parquet')
    df = df.merge(df_sci, on=merge_keys, how='left')

    df_e5 = pd.read_parquet(f'../test_feature/{proc}_multilingual_e5_large.parquet')
    df = df.merge(df_e5, on=merge_keys, how='left')

    df_minilm = pd.read_parquet(f'../test_feature/{proc}_minilm.parquet')
    df = df.merge(df_minilm, on=merge_keys, how='left')

    df_deb = pd.read_parquet(f'../test_feature/{proc}_deberta.parquet')
    df = df.merge(df_deb, on=merge_keys, how='left')

    df_w2v = pd.read_parquet(f'../test_feature/{proc}_w2v.parquet')
    df = df.merge(df_w2v, on=merge_keys, how='left')

    df_oag = pd.read_parquet(f'../test_feature/{proc}_oag_bert.parquet')
    df = df.merge(df_oag, on=merge_keys, how='left')

    df_tfidf = pd.read_parquet(f'../test_feature/{proc}_tfidf.parquet')
    df = df.merge(df_tfidf, on=merge_keys, how='left')

    df_graph = pd.read_parquet(f'../test_feature/{proc}_graph.parquet')
    df = df.merge(df_graph, on=merge_keys, how='left')

    df_jaccard = pd.read_parquet(f'../test_feature/{proc}_jaccard.parquet')
    df = df.merge(df_jaccard, on=merge_keys, how='left')

    df_basic = pd.read_parquet(f'../test_feature/{proc}_basic.parquet')
    df = df.merge(df_basic, on=merge_keys, how='left')
    
    # label data
    label_col = [col for col in df.columns if '_label' in col]
    df[label_col] = df[label_col].replace(-1, 0)
    prefix_list = list(set([p.split('_')[0] for p in label_col]))
    for p in prefix_list:
        cols = [col for col in label_col if p in col]
        df[f'{p}_label_mean'] = df[cols].sum(axis=1)/len(cols)
        df[f'{p}_label_mean'] = df[f'{p}_label_mean'].apply(lambda x : 1 / (1 + np.exp(-x)))
    type_list = list(set([p.split('_')[1] for p in label_col]))
    for t in type_list:
        cols = [col for col in label_col if t in col]
        df[f'{t}_label_mean'] = df[cols].sum(axis=1)/len(cols)
        df[f'{t}_label_mean'] = df[f'{t}_label_mean'].apply(lambda x : 1 / (1 + np.exp(-x)))
    # pred data
    pred_col = [col for col in df.columns if '_pred' in col]
    suffix_col =  list(set([p.split('_')[-1] for p in pred_col]))     
    for p in prefix_list:
        for s in suffix_col:
            cols = [col for col in pred_col if p in col]
            cols = [col for col in cols if s in col]
            if len(cols)>1:
                df[f'{p}_pred_mean_{s}'] = df[cols].sum(axis=1)/len(cols)
                df[f'{p}_pred_mean_{s}'] = df[f'{p}_pred_mean_{s}'].apply(lambda x : 1 / (1 + np.exp(-x)))
    feature_col = [col for col in df.columns.tolist() if col not in merge_keys]    
    return df,feature_col
df_train = get_ids(df_train_master, 'train')
df_train, feature_col = get_feature(df_train, 'train')
df_test = get_ids(df_test_master, 'test')
df_test, _ = get_feature(df_test, 'test')
print(f'train:{df_train[feature_col].shape} | test:{df_test[feature_col].shape}')

# 無意味な特徴量の削除
def drop_noise(df, f_col):
    one_value_cols = [col for col in f_col if df[col].nunique() <= 1]
    many_null_cols = [col for col in f_col if df[col].isnull().sum() / df_train.shape[0] > 0.9]
    noise_cols = list(set(many_null_cols + one_value_cols))
    return noise_cols
train_noise_cols = drop_noise(df_train, feature_col)
test_noise_cols = drop_noise(df_test, feature_col)
noise_cols = list(set(train_noise_cols + test_noise_cols))
before_num = len(feature_col)
feature_col = [col for col in feature_col if col not in noise_cols]
after_num = len(feature_col)
print(f'削減前:{before_num} | 削減後:{after_num}')
print(f'削除：{noise_cols}')

# Local Test

# 学習・検証・テストデータ作成
groups = df_train['author_id']
X = df_train.set_index(['author_id','paper_id'])
y = X['label']
X = X[feature_col]
# テストデータ作成
gss = GroupShuffleSplit(n_splits=1, test_size=0.3,random_state=85)
train_idxs,test_idxs = next(gss.split(X=X, y=y, groups=groups))
X, X_test, y, y_test = X.iloc[train_idxs],X.iloc[test_idxs], y.iloc[train_idxs],y.iloc[test_idxs]
# 学習・検証データ作成
X_train, X_valid, y_train, y_valid = train_test_split(
    X, 
    y, 
    test_size=0.1,
    random_state=1,
    shuffle=True,
    stratify=y
)
print(f'train:{X_train.shape} 負例:{len(y_train[y_train==0])} | valid:{X_valid.shape} 負例:{len(y_valid[y_valid==0])} | test:{X_test.shape} 負例:{len(y_test[y_test==0])}')

# 評価関数

# 評価関数
def get_evaluate(y_test, predict):
    auc = roc_auc_score(y_test, predict)
    predict = np.where(predict < 0.5, 0, 1)
    precision = precision_score(y_test, predict)
    recall = recall_score(y_test, predict)      
    return auc, precision, recall

def compute_score(target: np.ndarray,pred: np.ndarray,groups: np.ndarray) -> float:
    result_df = pd.DataFrame({"pred": pred, "target": target, "group": groups})
    total_errors = len(result_df[result_df["target"]==0])
    score = 0.0
    for _, gdf in result_df.groupby("group"):
        weight = len(gdf[gdf["target"]==0]) / total_errors
        auc = roc_auc_score(gdf["target"], gdf["pred"])
        score += auc * weight
    return score
    
def eval(y_pred, X_test, y_test):
    local_score = compute_score(y_test.values, np.array(y_pred), y_test.index.get_level_values('author_id').tolist())
    auc, precision, recall = get_evaluate(y_test, y_pred)
    print(f'LOCAL SCORE:{local_score}')
    print(f'AUC:{auc} | PRECISION:{precision} | RECALL:{recall}')
    
    y_pred = np.where(y_pred < 0.5, 0, 1)
    test = X_test.copy()
    test['pred'] = y_pred
    test['label'] = y_test.values
    test = test.reset_index()
    tmp = []
    for author_id, group in test.groupby('author_id'):
        group.loc[group['label'] == group['pred'], 'flag'] = 1
        group.loc[group['label'] != group['pred'], 'flag'] = 0
        error = (len(group[group['flag']==0])/len(group))*100
        group['error'] = error
        group = group[['author_id','paper_id','pred','label','flag','error']]
        tmp.append(group)
    test = pd.concat(tmp)
    test = test.merge(df_master,on='paper_id',how='left')
    test = test.sort_values('error',ascending=False)
    cm = confusion_matrix(y_test.values, y_pred)
    print(cm)
    return local_score, test

# モデル定義

def nn_train(n, X_train, X_valid, y_train, y_valid, X_test, y_test=None, log=False):    
    # 欠損値補完
    df_mean = X_train.mean(numeric_only=True)
    X_train_nn = X_train.fillna(df_mean)
    X_valid_nn = X_valid.fillna(df_mean)
    X_test_nn = X_test.fillna(df_mean)
    # 次元圧縮
    dim=512
    pca = PCA(n_components=dim)
    pca.fit(X_train_nn)
    X_train_nn = pca.transform(X_train_nn)
    X_valid_nn = pca.transform(X_valid_nn)
    X_test_nn = pca.transform(X_test_nn)
    # 標準化
    ss = StandardScaler()
    X_train_nn = ss.fit_transform(X_train_nn)
    X_valid_nn = ss.transform(X_valid_nn)
    X_test_nn = ss.transform(X_test_nn) 
    # モデルの構築
    model = Sequential([
        Dense(256, input_dim=X_train_nn.shape[1], activation='relu'),
        BatchNormalization(),
        Dropout(0.6),
        Dense(64, input_dim=X_train_nn.shape[1], activation='relu'),
        BatchNormalization(),
        Dropout(0.4),
        Dense(1, activation='sigmoid')
    ])
    if log:
        model.summary()
    early_stopping = EarlyStopping(monitor='val_auc', patience=10, mode='max', restore_best_weights=True)

    # モデルのコンパイル
    model.compile(optimizer=Adam(learning_rate=0.001),
                  loss=BinaryCrossentropy(),
                  metrics=[AUC(name='auc')])
    # モデルのトレーニング
    history = model.fit(X_train_nn, y_train,
                        epochs=1000,
                        batch_size=256,
                        validation_data=(X_valid_nn, y_valid),
                        verbose=10,
                        callbacks=[early_stopping]
                       )
    y_pred = model.predict(X_test_nn).flatten()
    del model
    torch.cuda.empty_cache()
    if y_test is not None:
        _, test = eval(y_pred, X_test, y_test)
        return y_pred
    else:
        return y_pred
nn_pred = nn_train(1, X_train, X_valid, y_train, y_valid, X_test, y_test, True)

def tab_train(n, X_train, X_valid, y_train, y_valid, X_test, y_test=None, log=False):
    # 事前学習
    # f_feature = [col for col in feature_col if col not in null_col]
    df_mean = X_train.mean(numeric_only=True)
    X_train_tab = X_train.fillna(df_mean)
    X_valid_tab = X_valid.fillna(df_mean)
    X_test_tab = X_test.fillna(df_mean)
    device = "cuda" if torch.cuda.is_available() else "cpu"
    unsupervised_model = TabNetPretrainer(
        optimizer_fn=torch.optim.Adam,
        optimizer_params=dict(lr=2e-2),
        device_name = device,
        mask_type='entmax',
        verbose = 5,
    )
    unsupervised_model.fit(
        X_train_tab.values,
        eval_set=[X_valid_tab.values],
        batch_size = 128,
        pretraining_ratio=0.8,
        # num_workers=CORE_NUM-1,
    )
    # 本学習
    model = TabNetClassifier(
        # n_d=16,
        # n_a=16,
        # n_steps=4, 
        optimizer_fn=torch.optim.Adam,
        optimizer_params=dict(lr=2e-2),
        device_name = device,
        verbose = 10,
        seed = n
    )
    model.fit(
        X_train_tab.values, y_train.values,
        eval_set = [(X_valid_tab.values, y_valid.values)],
        eval_metric= ['auc'],
        eval_name = ['valid'],
        batch_size = 128,
        max_epochs = 100,
        patience = 5,
    num_workers=CORE_NUM-1,
        from_unsupervised=unsupervised_model
    )
    y_pred = model.predict_proba(X_test_tab.values)[:, 1]
    if y_test is not None:
        _, test = eval(y_pred, X_test, y_test)
        return y_pred
    else:
        return y_pred

tab_pred = tab_train(1, X_train, X_valid, y_train, y_valid, X_test, y_test, False)

def rf_train(n, X_train, X_valid, y_train, y_valid, X_test, y_test=None, log=False):
    print(f'TRAIN SHAPE:{X_train.shape}')
    random_forest_model = RandomForestClassifier(n_estimators=150,random_state=42, max_depth=10)
    random_forest_model.fit(X_train[feature_col], y_train)
    y_pred = random_forest_model.predict_proba(X_test[feature_col])[:, 1]
    if y_test is not None:
        _, test = eval(y_pred, X_test, y_test)
        return y_pred
    else:
        return y_pred
rf_pred = rf_train(1, X_train, X_valid, y_train, y_valid, X_test, y_test, False)

def lgbm_train(n, X_train, X_valid, y_train, y_valid, X_test, y_test=None, log=False):
    print(f'TRAIN SHAPE:{X_train.shape}')
    lgb_train = lgb.Dataset(
        X_train, 
        y_train
    )
    lgb_eval = lgb.Dataset(
        X_valid, 
        y_valid, 
        reference=lgb_train,
    )
    params = {
        'objective': 'binary',
        'metric': 'auc',
        'boosting_type': 'gbdt',
        'max_depth': 7,
        'num_leaves': 26,
        'learning_rate': 0.02,
        'verbosity': -1,
        'random_state': n,
        'n_jobs':CORE_NUM - 1
    }
    model = lgb.train(
        params,
        lgb_train,
        valid_sets=[lgb_eval],
        num_boost_round=10000,
        callbacks=[
            lgb.early_stopping(
                stopping_rounds=10,
                first_metric_only=True, 
                verbose=True),
            lgb.log_evaluation(1000)
        ]
    )
    importance = pd.DataFrame(model.feature_importance(), index=X_train.columns.tolist(), columns=['importance'])
    importance = importance.sort_values('importance',ascending=False)
    if log:
        print(f'NUM_OF_PARAMS:{len(model.params)}')
        print(f'PARAMS:{model.params}')
        display(importance.head())
    y_pred = model.predict(X_test, num_iteration=model.best_iteration)
    if y_test is not None:
        _, test = eval(y_pred, X_test, y_test)
        return y_pred, importance, test, model.params, model
    else:
        return y_pred, importance

lgb_pred, lgb_imp, lgb_error, params, model = lgbm_train(1, X_train, X_valid, y_train, y_valid, X_test, y_test, False)

def catboost_train(n,X_train, X_valid, y_train, y_valid, X_test, y_test=None,log=False):
    train_pool = Pool(X_train, label=y_train)
    valid_pool = Pool(X_valid, label=y_valid)
    params = {
        # タスク設定と損失関数
        'loss_function': 'Logloss',
        'eval_metric':'AUC',
        'depth':7,
        'num_boost_round': 10000,
        'early_stopping_rounds': 10,
        'random_state':n,
        'thread_count':CORE_NUM-1,
        'use_best_model':True,
    }
    # モデルを学習する
    model = CatBoost(params)
    model.fit(
        train_pool,
        eval_set=valid_pool,
        verbose=1000,
    )
    importance = pd.DataFrame(
         model.get_feature_importance(type='PredictionValuesChange'),
         index=X_train.columns.tolist(),
         columns=['importance'])
    importance = importance.sort_values('importance',ascending=False)
    if log:
        print(f'NUM_OF_PARAMS:{len(model.get_params())}')
        print(f'PARAMS:{model.get_params()}')
        display(importance.head())
    y_pred = model.predict(X_test.values.tolist(),prediction_type='Probability')[:,1]
    if y_test is not None:
        _, test = eval(y_pred, X_test, y_test)
        return y_pred, importance, test
    else:
        return y_pred, importance
cat_pred, cat_imp, cat_error = catboost_train(1, X_train, X_valid, y_train, y_valid, X_test, y_test)

def xgbt_train(n,X_train, X_valid, y_train, y_valid, X_test, y_test=None,log=False):
    xgb_train = xgb.DMatrix(X_train,label=y_train)
    xgb_eval = xgb.DMatrix(X_valid,label=y_valid)
    xgb_test = xgb.DMatrix(X_test)
    xgb_params = {
        'objective': 'binary:logistic',
        'eval_metric': 'auc',
        'max_depth': 7,
        'max_leaves':26,
        'learning_rate': 0.02,
        'seed': n,
        # 'tree_method': 'gpu_hist',
    }
    model = xgb.train(
        xgb_params,
        xgb_train,
        num_boost_round=10000,
        evals=[(xgb_eval, 'eval')],
        early_stopping_rounds=10,
        verbose_eval=1000,
    )
    importance = pd.DataFrame(
        list(model.get_score(importance_type='gain').items()), 
        columns=['feature','importance']
    )
    importance = importance.sort_values('importance',ascending=False)
    if log:
        print(f'NUM_OF_PARAMS:{len(model.attributes())}')
        print(f'PARAMS:{model.attributes()}')
        display(importance.head())
    y_pred = model.predict(xgb_test,iteration_range=(0, model.best_iteration))
    if y_test is not None:
        _, test = eval(y_pred, X_test, y_test)
        return y_pred, importance, test
    else:
        return y_pred, importance
xgb_pred, xgb_imp, xgb_error = xgbt_train(1, X_train, X_valid, y_train, y_valid, X_test, y_test)

# ミックス
# lgb + cat
mean_pred = np.mean([lgb_pred, cat_pred], axis=0)
_, _ = eval(mean_pred, X_test, y_test)
# lgb + xgb
mean_pred = np.mean([lgb_pred, xgb_pred], axis=0)
_, _ = eval(mean_pred, X_test, y_test)
# xgb + cat
mean_pred = np.mean([cat_pred, xgb_pred], axis=0)
_, _ = eval(mean_pred, X_test, y_test)
# lgb + xgb + cat
mean_pred = np.mean([lgb_pred,cat_pred,xgb_pred], axis=0)
_, _ = eval(mean_pred, X_test, y_test)
# lgb + xgb + cat + rf
mean_pred = np.mean([lgb_pred, cat_pred, xgb_pred, rf_pred], axis=0)
_, _ = eval(mean_pred, X_test, y_test)
# lgb + xgb + cat + rf + tab
mean_pred = np.mean([lgb_pred, cat_pred, xgb_pred, rf_pred, tab_pred], axis=0)
_, _ = eval(mean_pred, X_test, y_test)
# lgb + xgb + cat + rf + tab + nn
mean_pred = np.mean([lgb_pred, cat_pred, xgb_pred, rf_pred, tab_pred, nn_pred], axis=0)
_, _ = eval(mean_pred, X_test, y_test)
# xgb + cat + rf
mean_pred = np.mean([lgb_pred, cat_pred, xgb_pred, rf_pred], axis=0)
_, _ = eval(mean_pred, X_test, y_test)
# xgb + cat + rf + tab
mean_pred = np.mean([cat_pred, xgb_pred, rf_pred, tab_pred], axis=0)
_, _ = eval(mean_pred, X_test, y_test)
# xgb + cat + rf + tab + nn
mean_pred = np.mean([cat_pred, xgb_pred, rf_pred, tab_pred, nn_pred], axis=0)
_, _ = eval(mean_pred, X_test, y_test)

# SUBMIT

i = 0
X = df_train[feature_col]
y = df_train['label']
kf = StratifiedKFold(n_splits=5, shuffle=True, random_state=92)
df_result = pd.DataFrame()
for train_index, valid_index in kf.split(X, y):
    
    X_train, y_train = X.iloc[train_index], y.iloc[train_index]
    X_valid, y_valid = X.iloc[valid_index], y.iloc[valid_index]
    submit = df_test.copy()
    
    lgb_pred, _ = lgbm_train(i,X_train, X_valid, y_train, y_valid, submit[feature_col])
    submit[f'pred_l_{i}'] = lgb_pred
    
    cat_pred, _ = catboost_train(i,X_train, X_valid, y_train, y_valid, submit[feature_col])
    submit[f'pred_c_{i}'] = cat_pred
    
    xgb_pred, _ = xgbt_train(i,X_train, X_valid, y_train, y_valid, submit[feature_col])
    submit[f'pred_x_{i}'] = xgb_pred

    nn_pred = nn_train(i, X_train, X_valid, y_train, y_valid, submit[feature_col])
    submit[f'pred_n_{i}'] = nn_pred

    rf_pred = rf_train(i,X_train, X_valid, y_train, y_valid, submit[feature_col])
    submit[f'pred_r_{i}'] = rf_pred

    tab_pred = tab_train(1, X_train, X_valid, y_train, y_valid, submit[feature_col])
    submit[f'pred_t_{i}'] = tab_pred
    
    submit = submit[[
        'author_id',
        'paper_id', 
        f'pred_n_{i}', 
        f'pred_t_{i}',
        f'pred_c_{i}',
        f'pred_l_{i}',
        f'pred_x_{i}', 
        f'pred_r_{i}'
    ]]
    
    if i == 0:
        df_result = submit
    else:
        df_result = df_result.merge(submit, on=['author_id','paper_id'], how='left')
    i += 1

df_result.head()

use_col = [col for col in df_result.columns if '_x_' not in col]
use_col

p_dic = {}
for name, group in df_result[use_col].groupby('author_id'):
    p_dic[f'{name}'] = dict(zip(group['paper_id'].tolist(), group.mean(axis=1, numeric_only=True).tolist()))

tf = open("../submit/submit.json", "w")
json.dump(p_dic, tf)
tf.close()
